+++ Sources/Utils.swift
@@ -48,11 +48,6 @@ extension Optional: BoxType {
   }
 }
 
-extension ImplicitlyUnwrappedOptional: BoxType {
-  var unboxed: Any? {
-    return self ?? nil
-  }
-}
 
 class Box<T> {
   var unboxed: T