+++ Tests/DipTests/DefinitionTests.swift
@@ -116,7 +116,6 @@ class DefinitionTests: XCTestCase {
     let def = Definition<Service, ()>(scope: .unique) { ServiceImp() as Service }
     
     XCTAssertTrue(def.implementingTypes.contains(where: { $0 == Service?.self }))
-    XCTAssertTrue(def.implementingTypes.contains(where: { $0 == Service!.self }))
   }
   
 }